# proto/gen/openapi/openapi.yaml
...
openapi: 3.0.3
info:
    title: ""
    version: 0.0.1
paths:
    /api/v1/ask/conversations:
        get:
            tags:
                - AskService
            operationId: AskService_ListAskConversations
            parameters:
                - name: limit
                  in: query
                  schema:
                    type: integer
                    format: int32
                - name: query
                  in: query
                  schema:
                    type: string
                - name: archived
                  in: query
                  description: When unset, the current (unarchived) conversations are returned.
                  schema:
                    type: boolean
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ListAskConversationsResponse'
                default:
                    description: Default error response
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Status'
        post:
            tags:
                - AskService
            operationId: AskService_CreateAskConversation
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/CreateAskConversationRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/AskConversationResponse'
                default:
                    description: Default error response
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Status'
    /api/v1/ask/conversations/{conversationId}:
        get:
            tags:
                - AskService
            operationId: AskService_GetAskConversation
            parameters:
                - name: conversationId
                  in: path
                  required: true
                  schema:
                    type: string
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/AskConversationResponse'
                default:
                    description: Default error response
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Status'
    /api/v1/ask/conversations/{conversationId}/messages:
        get:
            tags:
                - AskService
            operationId: AskService_ListAskMessages
            parameters:
                - name: conversationId
                  in: path
                  required: true
                  schema:
                    type: string
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ListAskMessagesResponse'
                default:
                    description: Default error response
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Status'
        post:
            tags:
                - AskService
            description: |-
                CreateAskMessage posts a question and returns the user message plus the
                 generated answer. This is the unary path; the REST surface also exposes a
                 streaming variant for token-by-token delivery.
            operationId: AskService_CreateAskMessage
            parameters:
                - name: conversationId
                  in: path
                  required: true
                  schema:
                    type: string
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/CreateAskMessageRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/CreateAskMessageResponse'
                default:
                    description: Default error response
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Status'
...
components:
    schemas:
...
        AskConversation:
            type: object
            properties:
                id:
                    type: string
                title:
                    type: string
                status:
                    type: string
                contextScope:
                    type: string
                    description: |-
                        context_scope is the default time window for grounding: "recent_7_days" |
                         "recent_30_days" | "all". Per-message scope/source may override it.
                headMessageId:
                    type: string
                    description: |-
                        head_message_id is the currently active leaf when forks (regenerations /
                         branches) exist.
                pinnedTime:
                    type: string
                    format: date-time
                archivedTime:
                    type: string
                    format: date-time
                createdTime:
                    type: string
                    format: date-time
                updatedTime:
                    type: string
                    format: date-time
                deletedTime:
                    type: string
                    format: date-time
            description: |-
                AskConversation is a single Ask thread (a ChatGPT-style conversation grounded
                 in the user's memos).
        AskConversationResponse:
            type: object
            properties:
                conversation:
                    $ref: '#/components/schemas/AskConversation'
        AskMessage:
            type: object
            properties:
                id:
                    type: string
                conversationId:
                    type: string
                role:
                    type: string
                    description: role is "user" | "assistant".
                content:
                    type: string
                parentId:
                    type: string
                    description: parent_id is the message this one follows (empty for the first turn).
                forkOfId:
                    type: string
                    description: |-
                        fork_of_id, when set, marks this as an alternative (regenerated) answer to
                         the same parent as the referenced message.
                status:
                    type: string
                    description: |-
                        status is the message lifecycle marker; persisted messages currently use
                         "complete".
                sourceRefs:
                    type: array
                    items:
                        $ref: '#/components/schemas/AskSourceRef'
                model:
                    type: string
                createdTime:
                    type: string
                    format: date-time
                updatedTime:
                    type: string
                    format: date-time
                deletedTime:
                    type: string
                    format: date-time
                promptVersion:
                    type: string
                    description: |-
                        prompt_version identifies the answer instructions used for assistant
                         messages. User and legacy messages leave it empty.
            description: |-
                AskMessage is one turn in a conversation. Branching and regeneration are
                 modeled as a tree: parent_id links a follow-up to the message it answers, and
                 fork_of_id links a regenerated answer to the answer it replaces.
        AskSourceRef:
            type: object
            properties:
                memoId:
                    type: string
                entryDate:
                    type: string
                excerpt:
                    type: string
                rank:
                    type: integer
                    format: int32
            description: |-
                AskSourceRef is a citation: a memo an answer was grounded in, so the user can
                 jump back to the original record.
...
        CreateAskConversationRequest:
            type: object
            properties:
                title:
                    type: string
                contextScope:
                    type: string
        CreateAskMessageRequest:
            type: object
            properties:
                conversationId:
                    type: string
                content:
                    type: string
                contextScope:
                    type: string
                    description: |-
                        context_scope overrides the conversation's time window for this turn:
                         "recent_7_days" | "recent_30_days" | "all".
                parentId:
                    type: string
                    description: |-
                        parent_id branches a follow-up off a specific message (empty = continue the
                         active head).
                forkOfId:
                    type: string
                    description: |-
                        fork_of_id regenerates an answer: the new assistant message becomes an
                         alternative to the referenced one under the same parent.
                sourceKind:
                    type: string
                    description: |-
                        source_kind selects what the answer is grounded in for this turn:
                         "records" (raw memos, default) | "memo_summary" | "summaries".
        CreateAskMessageResponse:
            type: object
            properties:
                messages:
                    type: array
                    items:
                        $ref: '#/components/schemas/AskMessage'
...
        GoogleProtobufAny:
            type: object
            properties:
                '@type':
                    type: string
                    description: The type of the serialized message.
            additionalProperties: true
            description: Contains an arbitrary serialized message along with a @type that describes the type of the serialized message.
...
        ListAskConversationsResponse:
            type: object
            properties:
                conversations:
                    type: array
                    items:
                        $ref: '#/components/schemas/AskConversation'
        ListAskMessagesResponse:
            type: object
            properties:
                messages:
                    type: array
                    items:
                        $ref: '#/components/schemas/AskMessage'
...
        Status:
            type: object
            properties:
                code:
                    type: integer
                    description: The status code, which should be an enum value of [google.rpc.Code][google.rpc.Code].
                    format: int32
                message:
                    type: string
                    description: A developer-facing error message, which should be in English. Any user-facing error message should be localized and sent in the [google.rpc.Status.details][google.rpc.Status.details] field, or localized by the client.
                details:
                    type: array
                    items:
                        $ref: '#/components/schemas/GoogleProtobufAny'
                    description: A list of messages that carry the error details.  There is a common set of message types for APIs to use.
            description: 'The `Status` type defines a logical error model that is suitable for different programming environments, including REST APIs and RPC APIs. It is used by [gRPC](https://github.com/grpc). Each `Status` message contains three pieces of data: error code, error message, and error details. You can find out more about this error model and how to work with it in the [API Design Guide](https://cloud.google.com/apis/design/errors).'
...